Move State to its own module.
